Fixed loop errors
Changed long modulus to 1 instead of zero, so that when it multiplies itself by ten in the loop, it is not constantly outputting zero.
Also changed the log10(n) on line 36 to log10(square) since the comment says to compute the number of digits of the square, not of n.